--- a/Engine/Invokable/Invokable.cs
+++ b/Engine/Invokable/Invokable.cs
@@ -4,7 +4,7 @@ namespace OpenTap
 {
     /// <summary> Action(T) IInvokable. </summary>
     /// <typeparam name="T"></typeparam>
-    class Invokable<T> : IInvokable<T>
+    readonly struct Invokable<T> : IInvokable<T>
     {
         readonly Action<T> action;
         public Invokable(Action<T> action) => this.action = action;
@@ -14,12 +14,13 @@ namespace OpenTap
         /// <summary> Add an ignored argument. </summary>
         public Invokable<T, T2> AddArg<T2>()
         {
-            return new Invokable<T, T2>((a1, a2) => action(a1));
+            var action2 = action;
+            return new Invokable<T, T2>((a1, a2) => action2(a1));
         }
     }
     
     /// <summary> Action(T) IInvokable. </summary>
-    class Invokable<T, T2> : IInvokable<T,T2>
+    readonly struct Invokable<T, T2> : IInvokable<T,T2>
     {
         readonly Action<T,T2> action;
         public Invokable(Action<T,T2> action) => this.action = action;