--- a/Engine/TestStepRun.cs
+++ b/Engine/TestStepRun.cs
@@ -262,30 +262,33 @@ namespace OpenTap
             completedEvent.Set();
         }
 
+        
+        ITypeData stepTypeData;
+
+        TestStepRun(ITestStep step)
+        {
+            TestStepId = step.Id;
+            TestStepName = step.GetFormattedName();
+            stepTypeData = TypeData.GetTypeData(step);
+            TestStepTypeName = stepTypeData.AsTypeData().AssemblyQualifiedName;
+            Parameters = ResultParameters.GetParams(step);
+            Verdict = Verdict.NotSet;
+        }
+        
         /// <summary>
         /// Constructor for TestStepRun.
         /// </summary>
         /// <param name="step">Property Step.</param>
         /// <param name="parent">Property Parent. </param>
         /// <param name="attachedParameters">Parameters that will be stored together with the actual parameters of the steps.</param>
-        public TestStepRun(ITestStep step, Guid parent, IEnumerable<ResultParameter> attachedParameters = null)
+        public TestStepRun(ITestStep step, Guid parent, IEnumerable<ResultParameter> attachedParameters = null): this(step)
         {
-            TestStepId = step.Id;
-            TestStepName = step.GetFormattedName();
-            TestStepTypeName = TypeData.FromType(step.GetType()).AssemblyQualifiedName;
-            Parameters = ResultParameters.GetParams(step);
-            Verdict = Verdict.NotSet;
             if (attachedParameters != null) Parameters.AddRange(attachedParameters);
             Parent = parent;
         }
         
-        internal TestStepRun(ITestStep step, TestRun parent, IEnumerable<ResultParameter> attachedParameters = null)
+        internal TestStepRun(ITestStep step, TestRun parent, IEnumerable<ResultParameter> attachedParameters = null): this(step)
         {
-            TestStepId = step.Id;
-            TestStepName = step.GetFormattedName();
-            TestStepTypeName = TypeData.FromType(step.GetType()).AssemblyQualifiedName;
-            Parameters = ResultParameters.GetParams(step);
-            Verdict = Verdict.NotSet;
             if (attachedParameters != null) Parameters.AddRange(attachedParameters);
             Parent = parent.Id;
             BreakCondition = calculateBreakCondition(step, parent);
@@ -369,13 +372,17 @@ namespace OpenTap
         ManualResetEventSlim runDone = new ManualResetEventSlim(false, 0);
         ManualResetEventSlim deferDone = new ManualResetEventSlim(false, 0);
         
-        internal void AfterRun(ITestStep step)
+        internal void AfterRun(ITestStep step, TestPlanRun testPlanRun)
         {
-            var resultMembers = TypeData.GetTypeData(step)
+            var resultMembers = stepTypeData
                 .GetMembers()
                 .Where(member => member.HasAttribute<ResultAttribute>())
                 .ToArray();
 
+            if (resultMembers.Length > 0 && ResultSource == null)
+            {
+                ResultSource = new ResultSource(this, testPlanRun);
+            }
             void publishResults()
             {
                 foreach (var r in resultMembers)
@@ -399,13 +406,13 @@ namespace OpenTap
                 ResultSource.Defer(() =>
                 {
                     deferDone.Set();
-                    stepRuns.Clear();
+                    stepRuns = null;
                 });
             }
             else
             {
                 deferDone.Set();
-                stepRuns.Clear();
+                stepRuns = null;
             }
         }
 
@@ -419,7 +426,7 @@ namespace OpenTap
         /// <exception cref="TimeoutException"></exception>
         internal TestStepRun WaitForChildStepStart(Guid childStep, int timeout, bool wait)
         {
-            if (stepRuns.TryGetValue(childStep, out var run))
+            if (stepRuns != null && stepRuns.TryGetValue(childStep, out var run))
                 return run;
             if (!wait) return null;
             if (isCompleted) return null;