… Shared/ReflectionHelper.cs