Mostly these are focused around reducing GC overhead.

- Added option to reduce logging using EngineSettings
- Invokabe class turned into struct
- Made TestStep.Results lazy, so that it is only created if the step actually needs result processing.
- TestStepRun now keeps a field with the step type to avoid getting it later.

Measurements:

Two images installed into each folders:
```
tap.exe image install -t ./test "OpenTAP:9.19.1,OpenTap.Benchmark:beta"
tap.exe image install -t ./test2 "OpenTAP:9.19.1-alpha.6.1+de732a16.pull-874-merge,OpenTap.Benchmark:beta"
```
Without optimizations: 
```
Time spent per iteration: 4.59 ms (σ: 0.69 ms)
Benchmark completed 1000 iterations in 00:00:05.1020834
Time spent per iteration: 4.57 ms (σ: 0.66 ms)
Benchmark completed 1000 iterations in 00:00:04.9561905
Time spent per iteration: 4.55 ms (σ: 0.59 ms)
Benchmark completed 1000 iterations in 00:00:04.9333377
```

With Optimizations and logging

```
Time spent per iteration: 4.2 ms (σ: 0.45 ms)
Benchmark completed 1000 iterations in 00:00:04.5623400

Time spent per iteration: 4.35 ms (σ: 0.45 ms)
Benchmark completed 1000 iterations in 00:00:04.6975976

Time spent per iteration: 4.23 ms (σ: 0.44 ms)
Benchmark completed 1000 iterations in 00:00:04.6370493
```

With reduced logging:

```
Time spent per iteration: 3.1 ms (σ: 0.33 ms)
Benchmark completed 1000 iterations in 00:00:03.2975280
Time spent per iteration: 3.14 ms (σ: 0.33 ms)
Benchmark completed 1000 iterations in 00:00:03.3215842
Time spent per iteration: 3.09 ms (σ: 0.31 ms)
```

Update 2: After further improvements (commit 2)
```
Time spent per iteration: 2.24 ms (σ: 0.26 ms)
Benchmark completed 1000 iterations in 00:00:02.3327779
Time spent per iteration: 2.13 ms (σ: 0.24 ms)
Benchmark completed 1000 iterations in 00:00:02.2252589
Time spent per iteration: 2.17 ms (σ: 0.25 ms)
Benchmark completed 1000 iterations in 00:00:02.2431797
```